pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "gunicorn>=21.0",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
    "requests>=2.31",
    "validators>=0.22",
]

[dependency-groups]
dev = [
    "flake8>=7.0",
]